pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fast-agent"
version = "0.1.0"
description = "Lightning-fast agent for 4-second responses"
requires-python = ">=3.11"
dependencies = [
    # Anthropic - direct client (no SDK overhead)
    "anthropic>=0.40.0",

    # API
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.10.0",

    # Storage - direct clients
    "supabase>=2.10.0",
    "neo4j>=5.26.0",

    # Embeddings
    "openai>=1.55.0",

    # LLM
    "google-generativeai>=0.8.0",

    # Utils
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
